+++ src/cogs/tasks.py
@@ -28,7 +28,6 @@ class tasks(commands.Cog):
         self.update_users.start()
         self.discord_status.start()
         # self.remove_records_from_previous_day.start()
-        self.new_github_release.start()
         self.free_games.start()
         self.joined_date.start()
         log.info("Waiting for tasks to start")
@@ -59,38 +58,6 @@ class tasks(commands.Cog):
                     await self.bot.channel_send(
                         guild, channel_type="games", msg="a", embed=embed
                     )
-
-    @tasks.loop(hours=1)
-    async def new_github_release(self):
-        "Checks if there is a new release on github and sends a message to the channel"
-        r = requests.request(
-            "GET", "https://api.github.com/repos/tekofx/furbot/releases"
-        )
-        r = r.json()
-        for guild in self.bot.guilds:
-            if not exists_channel(guild, "bot_news"):
-                continue
-
-            if not check_record_in_database(guild, r[0]["url"]):
-                create_record(guild, ["github", r[0]["url"]])
-                version = r[0]["tag_name"]
-                release_changelog = r[0]["body"]
-                url = r[0]["html_url"]
-                embed = nextcord.Embed(
-                    title="Nueva versión " + version, description=url
-                )
-                embed.set_thumbnail(
-                    url="https://raw.githubusercontent.com/tekofx/furbot/main/assets/furbot_logo.png"
-                )
-                release_changelog = release_changelog.replace("\r", "")
-                release_changelog = release_changelog.split("#")
-                for i in release_changelog:
-                    if i != "":
-                        i = i.splitlines()
-                        var = "\n".join(i[1:])
-                        embed.add_field(name=i[0], value=var, inline=False)
-
-                await self.bot.channel_send(guild, "bot_news", "a", embed)
 
     @tasks.loop(hours=30)
     async def remove_records_from_previous_day(self):