+++ Changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## v4.3.6 - 2022-5-10
+### Arreglado
+- Hora de activacion de task meme
+
+### Modificado
+- Comprobacion de nueva version al iniciar bot
+
 
 ## v4.3.5 - 2022-5-10
 ### Añadido